Add settime and time commands for system time management
Implemented the `settime` command to allow users to set the system time using a Unix timestamp, along with the `time` command to display the current local and UTC time. Enhanced the NTP synchronization process to send the synced time to a connected peer if dual communication is established. Updated command registration to include the new functionalities.